# docs/openapi.yaml
openapi: 3.0.0
info:
  title: CapeControl Authentication API
  description: |
    Secure, scalable authentication system for CapeControl with JWT tokens, 
    role-based access control, and developer revenue tracking.
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (Customer, Developer, Admin)
    - Secure password reset workflow
    - Developer earnings and revenue tracking
    - Comprehensive audit logging
    
    ## Security
    - bcrypt password hashing
    - HS256 JWT tokens with configurable expiration
    - Token revocation and session management
    - Rate limiting and abuse protection
    
  version: 1.0.0
  contact:
    name: CapeControl API Support
    email: api-support@capecontrol.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.capecontrol.com
    description: Production server
  - url: https://staging-api.capecontrol.com
    description: Staging server
  - url: http://localhost:8000
    description: Development server
security:
  - BearerAuth: []
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
            examples:
              customer:
                summary: Customer registration
                value:
                  email: "customer@example.com"
                  password: "SecurePassword123!"
                  first_name: "John"
                  last_name: "Doe"
                  role: "customer"
                  experience: "intermediate"
              developer:
                summary: Developer registration
                value:
                  email: "developer@example.com"
                  password: "DevPassword123!"
                  first_name: "Jane"
                  last_name: "Smith"
                  role: "developer"
                  company: "TechCorp"
                  website: "https://techcorp.com"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Login with email and password to receive JWT tokens
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "SecurePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or refresh endpoint
  schemas:
    UserCreateRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "SecurePassword123!"
          description: "Minimum 8 characters, must include uppercase, lowercase, and number"
        first_name:
          type: string
          maxLength: 100
          example: "John"
        last_name:
          type: string
          maxLength: 100
          example: "Doe"
        role:
          type: string
          enum: [customer, developer, admin]
          default: customer
          example: "customer"
        phone:
          type: string
          maxLength: 20
          example: "+1-555-0123"
        website:
          type: string
          format: uri
          example: "https://example.com"
        company:
          type: string
          maxLength: 255
          example: "Example Corp"
        experience:
          type: string
          enum: [beginner, intermediate, advanced, expert]
          example: "intermediate"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "SecurePassword123!"
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "user@example.com"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        role:
          type: string
          enum: [customer, developer, admin]
          example: "customer"
        phone:
          type: string
          example: "+1-555-0123"
        website:
          type: string
          example: "https://example.com"
        company:
          type: string
          example: "Example Corp"
        experience:
          type: string
          enum: [beginner, intermediate, advanced, expert]
          example: "intermediate"
        is_active:
          type: boolean
          example: true
        is_verified:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2025-01-08T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-08T11:00:00Z"
        last_login_at:
          type: string
          format: date-time
          example: "2025-01-08T10:30:00Z"
    Tokens:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 3600
          description: "Access token expiration time in seconds"
...
    AuthSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "User registered successfully"
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tokens:
              $ref: '#/components/schemas/Tokens'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error_code:
          type: string
          example: "INVALID_CREDENTIALS"
        details:
          type: object
          additionalProperties: true
          example:
            field: "Additional error details"
...